Safari-611.1.10/LayoutTests/media/video-initially-hidden-volume-slider-up-expected.txt


EXPECTED (volumeSliderCoordinates[0] <= muteButtonCoordinates[0] && volumeSliderCoordinates[1] <= muteButtonCoordinates[1] == 'true') OK
END OF TEST
